like코드 라우터, 컨트롤러 분리
